[JDC-42-Rename-Skill-Chains] Rename and remove calculatorBase file
