+++ src/calculator/chains/enchant-hand.ts
@@ -4,7 +4,7 @@ import {
   increasedSansheng,
   insertPayloadSkills,
   makeSkillBaseInformation,
-} from '@calculator/skillChain';
+} from '@calculator/skill-chains';
 import ChainComponent from '@componet/chain';
 import { pipe } from '@componet/compose';
 import { createSkillFactory } from '@packages/core/skill';
