+++ src/calculator/chains/heng-sao-liu-he-dot.ts
@@ -5,7 +5,7 @@ import {
   increasedSongYan,
   insertPayloadSkills,
   makeSkillBaseInformation,
-} from '@calculator/skillChain';
+} from '@calculator/skill-chains';
 import ChainComponent from '@componet/chain';
 import { pipe } from '@componet/compose';
 import { createSkillFactory } from '@packages/core/skill';
