+++ src/calculator/chains/qian-jin-zhui.ts
@@ -3,7 +3,7 @@ import {
   increasedSansheng,
   insertPayloadSkills,
   makeSkillBaseInformation,
-} from '@calculator/skillChain';
+} from '@calculator/skill-chains';
 import ChainComponent from '@componet/chain';
 import { pipe } from '@componet/compose';
 import { isJinGangRiLunEnchat } from '@componet/utils';
