+++ src/calculator/chains/shou-que-shi.ts
@@ -5,7 +5,7 @@ import {
   increasedSansheng,
   insertPayloadSkills,
   makeSkillBaseInformation,
-} from '@calculator/skillChain';
+} from '@calculator/skill-chains';
 import ChainComponent from '@componet/chain';
 import { pipe } from '@componet/compose';
 import { increaseCriticalDamagePower, increaseSolarCriticalStrike } from '@componet/utils';
