+++ src/calculator/chains/surplus-value.ts
@@ -3,7 +3,7 @@ import {
   getDamageCoefficientParams,
   insertPayloadSkills,
   makeSkillBaseInformation,
-} from '@calculator/skillChain';
+} from '@calculator/skill-chains';
 import ChainComponent from '@componet/chain';
 import { createSkillFactory } from '@packages/core/skill';
 import { SkillNames } from '@types';
