+++ src/calculator/chains/ti-hu-guan-ding.ts
@@ -4,7 +4,7 @@ import {
   increasedSansheng,
   insertPayloadSkills,
   makeSkillBaseInformation,
-} from '@calculator/skillChain';
+} from '@calculator/skill-chains';
 import ChainComponent from '@componet/chain';
 import { pipe } from '@componet/compose';
 import { isXinZhengVersion } from '@componet/utils';
