+++ src/calculator/chains/wei-tuo-xian-chu.ts
@@ -6,7 +6,7 @@ import {
   increasedShenYi,
   insertPayloadSkills,
   makeSkillBaseInformation,
-} from '@calculator/skillChain';
+} from '@calculator/skill-chains';
 import ChainComponent from '@componet/chain';
 import { pipe } from '@componet/compose';
 import { deepClone, increaseSolarCriticalStrike } from '@componet/utils';
